\name{version.ncdf}
\alias{version.ncdf}
\title{Report version of ncdf package}
\description{
 Returns a string that is the version number of the ncdf package.
}
\usage{
 version.ncdf()
}
\value{
 A string (not float) that is the version number of the ncdf package.
}
\references{
 http://www.unidata.ucar.edu/packages/netcdf/
}
\details{
 Note that the returned value it is a string, not a floating point number.
}
\author{David W. Pierce \email{dpierce@ucsd.edu}}
\keyword{utilities}
