\name{ncdf-internal}
\title{Internal ncdf functions}
\alias{dim.create.ncdf}
\alias{dim.same.ncdf}
\alias{default.missval.ncdf}
\alias{in.list.name.ncdf}
\alias{varname.inq.ncd}
\alias{vobjtovarid}
\alias{var.inq.ncdf}
\alias{dim.inq.ncdf}
\alias{unlimdim.ncdf}
\alias{varsize.ncdf}
\alias{varndims.ncdf}
\alias{dimid.inq.ncdf}
\alias{varid.inq.ncdf}
\alias{varid.to.missing.value}
\alias{varname.inq.ncdf}
\alias{vartype.ncdf}
\alias{vobjtodimid}
\alias{vobjtodimname}
\alias{blankstring.ncdf}
\description{
 Internal ncdf functions.
}
\details{
 These are not to be called by the user.
}
\keyword{internal}

